+++ mui/demos/showcase/.eslintrc.js
@@ -1,0 +1,11 @@
+module.exports =  {
+        root: true,
+        parser:  '@typescript-eslint/parser',
+        extends:  [ '@pxblue/eslint-config/tsx' ],
+        parserOptions:  {
+            project: "./tsconfig.json",
+        },
+        env: {
+            browser: true
+        }
+    };